backend/requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-CORS==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0

bcrypt==4.0.1
pyjwt==2.8.0
requests==2.31.0
python-dateutil==2.8.2
gunicorn==21.2.0
flask-limiter==3.5.0
marshmallow==3.20.1
requests==2.31.0
flask-socketio==5.3.2